feat: add useRoom hooks and refactor AccFin
