+++ src/components/edit/student/AccFinStu.tsx
@@ -1,35 +1,15 @@
-import React, { FC, useEffect, useState } from 'react'
+import React, { FC } from 'react'
 
 import { HotTable, HotTableProps } from '@handsontable/react-wrapper';
 import 'handsontable/styles/handsontable.css';
 import 'handsontable/styles/ht-theme-main.css';
-import { Button, message, Popconfirm } from 'antd';
+import { Button, Popconfirm } from 'antd';
 import { useStudentGrid } from '@/hooks/useStudentGrid';
+import { useRoom } from '@/hooks/useRoom';
 
 const AccFinStu : FC<HotTableProps> = () => {
   const { hotRef, gridRows, isSaving, handleSave, handleAdd, studentTypes } = useStudentGrid('AccFin');
-  const [roomOptions, setRoomOptions] = useState<string[]>([]);
-
-
-  useEffect(() => {
-    fetch('/api/room?location=AccountingFinanceLevel')
-      .then((res) => {
-        if (!res.ok) throw new Error(`HTTP ${res.status}`);
-        return res.json();
-      })
-      .then((rooms: Array<{ roomNo?: string | null }>) => {
-        const options = Array.isArray(rooms)
-          ? rooms
-              .map((r) => (r.roomNo ?? '').toString().trim())
-              .filter((v) => v.length > 0)
-          : [];
-        setRoomOptions(options);
-      })
-      .catch((err) => {
-        console.error(err);
-        message.error('Failed to load rooms');
-      });
-  }, []);
+  const { roomOptions } = useRoom('AccountingFinanceLevel');
 
   return (
     <div className=' p-4'>
